New: Core user CLI
