+++ .gitignore
@@ -1,3 +1,6 @@
 node_modules/
+
 index.js
 blockchain.js
+chainlog.js
+usercli.js
