+++ README.md
@@ -4,9 +4,23 @@ A simple blockchain and cryptocurrency wallet implemented in Node.js and TypeScr
 
 # Usage
 
+Start cloning Blockchain:
 ```
 git clone https://github.com/staxhinho/Blockchain.git
+```
 
+
+Next, install its dependencies (make sure npm is installed on your machine):
+```
 npm install
+```
+
+To compile, run:
+```
+npm run dev
+```
+
+To start, run:
+```
 npm start
 ```
