+++ chainlog.ts
@@ -1,0 +1,3 @@
+import { Chain } from "./blockchain";
+
+console.log(Chain.instance);
