+++ package.json
@@ -15,5 +15,8 @@
   "devDependencies": {
     "@types/node": "^22.13.10",
     "typescript": "^5.8.2"
+  },
+  "dependencies": {
+    "inquirer": "^12.4.3"
   }
 }
